implement ImageMetadataListParser and ImageMetadataParser

+++ src/main/java/de/uni_passau/fim/se2/litterbox/ast/model/metadata/ressources/ImageMetadata.java
@@ -3,11 +3,11 @@ package de.uni_passau.fim.se2.litterbox.ast.model.metadata.ressources;
 import de.uni_passau.fim.se2.litterbox.ast.visitor.ScratchVisitor;
 
 public class ImageMetadata extends ResourceMetadata {
-    private int bitmapResolution;
+    private double bitmapResolution;
     private double rotationCenterX;
     private double rotationCenterY;
 
-    public ImageMetadata(String assetId, String name, String md5ext, String dataFormat, int bitmapResolution,
+    public ImageMetadata(String assetId, String name, String md5ext, String dataFormat, double bitmapResolution,
                          double rotationCenterX, double rotationCenterY) {
         super(assetId, name, md5ext, dataFormat);
         this.bitmapResolution = bitmapResolution;
@@ -15,7 +15,7 @@ public class ImageMetadata extends ResourceMetadata {
         this.rotationCenterY = rotationCenterY;
     }
 
-    public int getBitmapResolution() {
+    public double getBitmapResolution() {
         return bitmapResolution;
     }
